# kyc-service/internal/interfaces/http/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for KYCStatusResponseStatus.
const (
	Rejected  KYCStatusResponseStatus = "Rejected"
	Submitted KYCStatusResponseStatus = "Submitted"
	Verified  KYCStatusResponseStatus = "Verified"
)

// Valid indicates whether the value is a known member of the KYCStatusResponseStatus enum.
func (e KYCStatusResponseStatus) Valid() bool {
	switch e {
	case Rejected:
		return true
	case Submitted:
		return true
	case Verified:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// KYCStatusResponse defines model for KYCStatusResponse.
type KYCStatusResponse struct {
	CustomerId openapi_types.UUID `json:"customer_id"`

	// Reason Rejection reason (only present when status is Rejected)
	Reason         *string                 `json:"reason,omitempty"`
	Status         KYCStatusResponseStatus `json:"status"`
	VerificationId openapi_types.UUID      `json:"verification_id"`
}

// KYCStatusResponseStatus defines model for KYCStatusResponse.Status.
type KYCStatusResponseStatus string

// RejectKYCRequest defines model for RejectKYCRequest.
type RejectKYCRequest struct {
	Reason string `json:"reason"`
}

// SubmitKYCRequest defines model for SubmitKYCRequest.
type SubmitKYCRequest struct {
	// CustomerId Customer UUID v7
	CustomerId openapi_types.UUID `json:"customer_id"`
}

// SubmitKYCResponse defines model for SubmitKYCResponse.
type SubmitKYCResponse struct {
	VerificationId openapi_types.UUID `json:"verification_id"`
}

// VerificationID defines model for VerificationID.
type VerificationID = openapi_types.UUID

// InternalError defines model for InternalError.
type InternalError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// UnprocessableEntity defines model for UnprocessableEntity.
type UnprocessableEntity = Error

// SubmitKYCJSONRequestBody defines body for SubmitKYC for application/json ContentType.
type SubmitKYCJSONRequestBody = SubmitKYCRequest

// RejectKYCJSONRequestBody defines body for RejectKYC for application/json ContentType.
type RejectKYCJSONRequestBody = RejectKYCRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Submit a KYC verification request
	// (POST /kyc)
	SubmitKYC(w http.ResponseWriter, r *http.Request)
	// Get verification status
	// (GET /kyc/{id})
	GetKYCStatus(w http.ResponseWriter, r *http.Request, id VerificationID)
	// Approve a KYC verification (KYC passed)
	// (POST /kyc/{id}/approve)
	ApproveKYC(w http.ResponseWriter, r *http.Request, id VerificationID)
	// Reject a KYC verification (KYC failed)
	// (POST /kyc/{id}/reject)
	RejectKYC(w http.ResponseWriter, r *http.Request, id VerificationID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Submit a KYC verification request
// (POST /kyc)
func (_ Unimplemented) SubmitKYC(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get verification status
// (GET /kyc/{id})
func (_ Unimplemented) GetKYCStatus(w http.ResponseWriter, r *http.Request, id VerificationID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Approve a KYC verification (KYC passed)
// (POST /kyc/{id}/approve)
func (_ Unimplemented) ApproveKYC(w http.ResponseWriter, r *http.Request, id VerificationID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Reject a KYC verification (KYC failed)
// (POST /kyc/{id}/reject)
func (_ Unimplemented) RejectKYC(w http.ResponseWriter, r *http.Request, id VerificationID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// SubmitKYC operation middleware
func (siw *ServerInterfaceWrapper) SubmitKYC(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SubmitKYC(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetKYCStatus operation middleware
func (siw *ServerInterfaceWrapper) GetKYCStatus(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id VerificationID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetKYCStatus(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ApproveKYC operation middleware
func (siw *ServerInterfaceWrapper) ApproveKYC(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id VerificationID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApproveKYC(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RejectKYC operation middleware
func (siw *ServerInterfaceWrapper) RejectKYC(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id VerificationID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RejectKYC(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/kyc", wrapper.SubmitKYC)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/kyc/{id}", wrapper.GetKYCStatus)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/kyc/{id}/approve", wrapper.ApproveKYC)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/kyc/{id}/reject", wrapper.RejectKYC)
	})

	return r
}

type InternalErrorJSONResponse Error

type NotFoundJSONResponse Error

type UnprocessableEntityJSONResponse Error

type SubmitKYCRequestObject struct {
	Body *SubmitKYCJSONRequestBody
}

type SubmitKYCResponseObject interface {
	VisitSubmitKYCResponse(w http.ResponseWriter) error
}

type SubmitKYC201JSONResponse SubmitKYCResponse

func (response SubmitKYC201JSONResponse) VisitSubmitKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type SubmitKYC422JSONResponse struct {
	UnprocessableEntityJSONResponse
}

func (response SubmitKYC422JSONResponse) VisitSubmitKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type SubmitKYC500JSONResponse struct{ InternalErrorJSONResponse }

func (response SubmitKYC500JSONResponse) VisitSubmitKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetKYCStatusRequestObject struct {
	Id VerificationID `json:"id"`
}

type GetKYCStatusResponseObject interface {
	VisitGetKYCStatusResponse(w http.ResponseWriter) error
}

type GetKYCStatus200JSONResponse KYCStatusResponse

func (response GetKYCStatus200JSONResponse) VisitGetKYCStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetKYCStatus404JSONResponse struct{ NotFoundJSONResponse }

func (response GetKYCStatus404JSONResponse) VisitGetKYCStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetKYCStatus500JSONResponse struct{ InternalErrorJSONResponse }

func (response GetKYCStatus500JSONResponse) VisitGetKYCStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ApproveKYCRequestObject struct {
	Id VerificationID `json:"id"`
}

type ApproveKYCResponseObject interface {
	VisitApproveKYCResponse(w http.ResponseWriter) error
}

type ApproveKYC204Response struct {
}

func (response ApproveKYC204Response) VisitApproveKYCResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type ApproveKYC404JSONResponse struct{ NotFoundJSONResponse }

func (response ApproveKYC404JSONResponse) VisitApproveKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type ApproveKYC422JSONResponse struct {
	UnprocessableEntityJSONResponse
}

func (response ApproveKYC422JSONResponse) VisitApproveKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type ApproveKYC500JSONResponse struct{ InternalErrorJSONResponse }

func (response ApproveKYC500JSONResponse) VisitApproveKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type RejectKYCRequestObject struct {
	Id   VerificationID `json:"id"`
	Body *RejectKYCJSONRequestBody
}

type RejectKYCResponseObject interface {
	VisitRejectKYCResponse(w http.ResponseWriter) error
}

type RejectKYC204Response struct {
}

func (response RejectKYC204Response) VisitRejectKYCResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type RejectKYC404JSONResponse struct{ NotFoundJSONResponse }

func (response RejectKYC404JSONResponse) VisitRejectKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type RejectKYC422JSONResponse struct {
	UnprocessableEntityJSONResponse
}

func (response RejectKYC422JSONResponse) VisitRejectKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type RejectKYC500JSONResponse struct{ InternalErrorJSONResponse }

func (response RejectKYC500JSONResponse) VisitRejectKYCResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Submit a KYC verification request
	// (POST /kyc)
	SubmitKYC(ctx context.Context, request SubmitKYCRequestObject) (SubmitKYCResponseObject, error)
	// Get verification status
	// (GET /kyc/{id})
	GetKYCStatus(ctx context.Context, request GetKYCStatusRequestObject) (GetKYCStatusResponseObject, error)
	// Approve a KYC verification (KYC passed)
	// (POST /kyc/{id}/approve)
	ApproveKYC(ctx context.Context, request ApproveKYCRequestObject) (ApproveKYCResponseObject, error)
	// Reject a KYC verification (KYC failed)
	// (POST /kyc/{id}/reject)
	RejectKYC(ctx context.Context, request RejectKYCRequestObject) (RejectKYCResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// SubmitKYC operation middleware
func (sh *strictHandler) SubmitKYC(w http.ResponseWriter, r *http.Request) {
	var request SubmitKYCRequestObject

	var body SubmitKYCJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.SubmitKYC(ctx, request.(SubmitKYCRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "SubmitKYC")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(SubmitKYCResponseObject); ok {
		if err := validResponse.VisitSubmitKYCResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetKYCStatus operation middleware
func (sh *strictHandler) GetKYCStatus(w http.ResponseWriter, r *http.Request, id VerificationID) {
	var request GetKYCStatusRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetKYCStatus(ctx, request.(GetKYCStatusRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetKYCStatus")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetKYCStatusResponseObject); ok {
		if err := validResponse.VisitGetKYCStatusResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ApproveKYC operation middleware
func (sh *strictHandler) ApproveKYC(w http.ResponseWriter, r *http.Request, id VerificationID) {
	var request ApproveKYCRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ApproveKYC(ctx, request.(ApproveKYCRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ApproveKYC")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ApproveKYCResponseObject); ok {
		if err := validResponse.VisitApproveKYCResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// RejectKYC operation middleware
func (sh *strictHandler) RejectKYC(w http.ResponseWriter, r *http.Request, id VerificationID) {
	var request RejectKYCRequestObject

	request.Id = id

	var body RejectKYCJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RejectKYC(ctx, request.(RejectKYCRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RejectKYC")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RejectKYCResponseObject); ok {
		if err := validResponse.VisitRejectKYCResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}